allegedly working

+++ crates/sapflux-core/src/metadata_manifest.rs
@@ -142,7 +142,7 @@ pub async fn preflight_manifest(
             ));
         }
 
-        let exists = sqlx::query_scalar::<_, i64>(
+        let exists = sqlx::query_scalar::<_, i32>(
             r#"
                 SELECT 1
                 FROM deployments
